+++ m3cli/plugins/total-report.py
@@ -4,12 +4,14 @@ This logic is created to convert parameters from the Human readable format to
 appropriate for M3 SDK API request.
 """
 import json
+from m3cli.services.request_service import BaseRequest
 
 from m3cli.plugins.utils.plugin_utilities import processing_report_format
 from m3cli.utils.utilities import timestamp_to_iso
+from m3cli.plugins import parse_and_set_date_range
 
 
-def create_custom_request(request):
+def create_custom_request(request: BaseRequest) -> BaseRequest:
     """ Transform 'total-report' command parameters from the Human
     readable format to appropriate for M3 SDK API request.
 
@@ -18,14 +20,13 @@ def create_custom_request(request):
     :type request: BaseRequest
     """
     processing_report_format(request)
+    parse_and_set_date_range(request.parameters)
+
     params = request.parameters
-    from_date = params.get('from')
-    to_date = params.get('to')
-    if from_date >= to_date:
-        raise AssertionError('Parameter "from" can not be equal or greater '
-                             'than parameter "to"')
-    request.parameters['target'] = {'tenantGroup': params.pop('tenantGroup'),
-                                    'reportUnit': 'TENANT_GROUP'}
+    params['target'] = {
+        'tenantGroup': params.pop('tenantGroup'),
+        'reportUnit': 'TENANT_GROUP'
+    }
     target = params['target']
     if params.get('clouds'):
         target.update({
@@ -61,9 +62,11 @@ def create_custom_response(request, response):
                 each_row['billingPeriodEndDate'] = \
                     timestamp_to_iso(each_row.get('billingPeriodEndDate'))
             response_processed.append(each_row)
-        response_processed.append({'recordType': 'grandTotal',
-                                   'totalPrice': grand_total,
-                                   'currencyCode': 'USD'})
+        response_processed.append({
+            'recordType': 'grandTotal',
+            'totalPrice': grand_total,
+            'currencyCode': 'USD',
+        })
         return response_processed
     if response.get('message'):
         return response.get('message')