+++ tests/test_services/test_plugin_service/test_plugin_service.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, create_autospec
 
 from m3cli.services.plugin_service import INTEGRATION_SUFFIX_ATTRIBUTE_NAME
 from tests.test_services.test_plugin_service import TestPluginService
@@ -30,7 +30,8 @@ class TestPluginServiceApplyPlugin(TestPluginService):
         self.response_data = 'response_data'
         self.data = {
             'request': self.request_data,
-            'response': self.response_data
+            'response': self.response_data,
+            'view_type': 'test_view'
         }
         self.cmd_def = {
             'integration_request': True,
@@ -38,39 +39,44 @@ class TestPluginServiceApplyPlugin(TestPluginService):
         }
         self.init_service_object()
         self.plugin_return_value = 'return'
-        self.plugin_method = MagicMock(return_value=self.plugin_return_value)
-        self.service.validate_method = MagicMock(
-            return_value=self.plugin_method)
+
+        def real_plugin_method(request=None, response=None, view_type=None):
+            return self.plugin_return_value
+
+        self.plugin_method = create_autospec(
+            real_plugin_method, return_value=self.plugin_return_value
+        )
+
+        self.service.validate_method = MagicMock(return_value=self.plugin_method)
 
     def test_apply_plugin_not_method_type_request(self):
-        method_type = 'integration_request'
-        self.cmd_def = {
-            method_type: False
-        }
-        self.init_service_object()
-        result = self.service.apply_plugin(self.data, method_type)
+        self.service.cmd_def['integration_request'] = False
+        result = self.service.apply_plugin(self.data, 'integration_request')
         self.assertEqual(result, self.request_data)
 
     def test_apply_plugin_not_method_type_response(self):
-        method_type = 'integration_response'
-        self.cmd_def = {
-            method_type: False
-        }
-        self.init_service_object()
-        result = self.service.apply_plugin(self.data, method_type)
+        self.service.cmd_def['integration_response'] = False
+        result = self.service.apply_plugin(self.data, 'integration_response')
         self.assertEqual(result, self.response_data)
 
     def test_apply_plugin_success_request(self):
         result = self.service.apply_plugin(self.data, 'integration_request')
         self.service.validate_method.assert_called_with(
-            method='create_custom_request')
-        self.plugin_method.assert_called_with(self.request_data)
+            method='create_custom_request'
+        )
+        self.plugin_method.assert_called_with(
+            request=self.request_data, view_type='test_view'
+        )
         self.assertEqual(result, self.plugin_return_value)
 
     def test_apply_plugin_success_response(self):
         result = self.service.apply_plugin(self.data, 'integration_response')
         self.service.validate_method.assert_called_with(
-            method='create_custom_response')
-        self.plugin_method.assert_called_with(self.request_data,
-                                              self.response_data)
+            method='create_custom_response'
+        )
+        self.plugin_method.assert_called_with(
+            request=self.request_data,
+            response=self.response_data,
+            view_type='test_view'
+        )
         self.assertEqual(result, self.plugin_return_value)